requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.12.1
click==8.1.7
colorama==0.4.6
Django==5.0
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.0
factory-boy==3.3.0
Faker==21.0.0
hiredis==2.3.2
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
